requirements-dev.txt
setuptools>=61.0.0
numpy<2.0
onnx-weekly>=1.17.0.dev20240325
onnxruntime-gpu==1.16.3
typing_extensions
rich>=13.7.1
cupy-cuda11x==12.3

# Docs site
furo
jax[cpu]
matplotlib
myst-parser[linkify]
sphinx-copybutton
sphinx-exec-code
sphinx-gallery
sphinx>=6
myst_nb
chardet

# Torch lib
beartype!=0.16.0

# Testing
expecttest==0.1.6
hypothesis
parameterized
pytest-cov
pytest-randomly
pytest-subtests
pytest-xdist
pytest!=7.1.0
pyyaml
torch>=2.1
torchvision>=0.16.0
transformers>=4.37.2

# Lint
lintrunner>=0.10.7
lintrunner_adapters>=0.12.0
